design: Nav 컴포넌트 제작
